requirements.txt
langgraph>=0.0.10
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-groq>=0.0.1
tavily-python>=0.1.9
python-dotenv>=1.0.0
streamlit>=1.32.0
tenacity>=8.2.3
pytest>=8.0.0 